Feat: split admin zone and centcomm maps, new centcomm, shuttle refactor and fixes in away maps (#1656)

+++ code/game/objects/effects/decals/decal.dm
@@ -2,14 +2,24 @@
 	icon = 'icons/turf/decals.dmi'
 	icon_state = "warningline"
 	layer = TURF_DECAL_LAYER
+	var/do_not_delete_me = FALSE
 
 /obj/effect/turf_decal/Initialize(mapload)
 	. = ..()
-	qdel(src) // return INITIALIZE_HINT_QDEL <-- Doesn't work
+	if(!do_not_delete_me)
+		qdel(src) // return INITIALIZE_HINT_QDEL <-- Doesn't work
+	invisibility = INVISIBILITY_MAXIMUM
 
 /obj/effect/turf_decal/ComponentInitialize()
 	. = ..()
 	var/turf/T = loc
 	if(!istype(T)) //you know this will happen somehow
 		CRASH("Turf decal initialized in an object/nullspace")
+	T.AddComponent(/datum/component/decal, icon, icon_state, dir, CLEAN_GOD, color, null, null, alpha)
+
+/obj/effect/turf_decal/onShuttleMove()
+	. = ..()
+	var/turf/T = loc
+	if(!istype(T)) //you know this will happen somehow
+		return
 	T.AddComponent(/datum/component/decal, icon, icon_state, dir, CLEAN_GOD, color, null, null, alpha)
